requirements.txt
azure-functions

Flask
flask_sqlalchemy
flask-migrate
flask_login
python-dotenv
prettyprint